+++ flutter_inappwebview_platform_interface/lib/src/platform_webview_feature.dart
@@ -407,6 +407,10 @@ class WebViewFeature_ {
   static const WEBVIEW_MEDIA_INTEGRITY_API_STATUS =
       const WebViewFeature_._internal("WEBVIEW_MEDIA_INTEGRITY_API_STATUS");
 
+  ///Feature for [isFeatureSupported]. This feature covers [InAppWebViewSettings.backForwardCacheEnabled].
+  static const BACK_FORWARD_CACHE =
+      const WebViewFeature_._internal("BACK_FORWARD_CACHE");
+
   ///Feature for [isFeatureSupported]. This feature covers [WebMessagePort.postMessage] with `ArrayBuffer` type,
   ///[InAppWebViewController.postWebMessage] with `ArrayBuffer` type, and [JavaScriptReplyProxy.postMessage] with `ArrayBuffer` type.
   static const WEB_MESSAGE_ARRAY_BUFFER = const WebViewFeature_._internal(
@@ -672,6 +676,10 @@ class AndroidWebViewFeature_ {
   ///Feature for [isFeatureSupported]. This feature covers [InAppWebViewSettings.webViewMediaIntegrityApiStatus].
   static const WEBVIEW_MEDIA_INTEGRITY_API_STATUS =
       const AndroidWebViewFeature_._internal("WEBVIEW_MEDIA_INTEGRITY_API_STATUS");
+
+  ///Feature for [isFeatureSupported]. This feature covers [InAppWebViewSettings.backForwardCacheEnabled].
+  static const BACK_FORWARD_CACHE =
+      const AndroidWebViewFeature_._internal("BACK_FORWARD_CACHE");
 
   ///Feature for [isFeatureSupported]. This feature covers [InAppWebViewSettings.requestedWithHeaderMode].
   static const REQUESTED_WITH_HEADER_CONTROL =